restructure inspector controls
