+++ src/directory/controls/index.js
@@ -1,2 +1,1 @@
-export { default as OrderControls } from './order-controls';
 export { default as RoleControl } from './role';
