+++ src/directory/edit/inspector-controls/constants.js
@@ -1,10 +1,9 @@
 /**
  * WordPress dependencies
  */
-import { SelectControl } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 
-const orderByOptions = [
+export const orderByOptions = [
 	{
 		// translators: label for ordering users by name.
 		label: __( 'Display name', 'simple-user-listing' ),
@@ -22,7 +21,7 @@ const orderByOptions = [
 	},
 ];
 
-const orderOptions = [
+export const orderOptions = [
 	{
 		// translators: label for ordering users in ascending order.
 		label: __( 'Ascending, ex: A → Z', 'simple-user-listing' ),
@@ -34,28 +33,3 @@ const orderOptions = [
 		value: 'desc',
 	},
 ];
-export default function OrderControls( { order, orderBy, onOrderChange, onOrderByChange } ) {
-	return (
-		<>
-			<SelectControl
-				__nextHasNoMarginBottom
-				label={ __( 'Order by', 'simple-user-listing' ) }
-				value={ orderBy }
-				options={ orderByOptions }
-				onChange={ ( value ) => {
-					onOrderByChange( value );
-				} }
-			/>
-
-			<SelectControl
-				__nextHasNoMarginBottom
-				label={ __( 'Order', 'simple-user-listing' ) }
-				value={ order }
-				options={ orderOptions }
-				onChange={ ( value ) => {
-					onOrderChange( value );
-				} }
-			/>
-		</>
-	);
-}
